# A few micro-optimizations for ParseRawQuery