Testing against 2 longer queries just pulled from StackOverflow. One LINQ (~32% improvement) and one just a select with a few joins (~16% improvement).

**Before**
|      Method |       Mean |    Error |   StdDev |
|------------ |-----------:|---------:|---------:|
|   ParseLinq | 2,418.0 ns | 47.37 ns | 84.20 ns |
| ParseSelect |   897.8 ns | 16.00 ns | 14.18 ns |

**After**
|      Method |       Mean |    Error |   StdDev |
|------------ |-----------:|---------:|---------:|
|   ParseLinq | 1,627.3 ns | 18.17 ns | 14.19 ns |
| ParseSelect |   754.1 ns |  8.89 ns |  7.88 ns |

- One of the rarer cases where using spans just for the StringBuilder cases did not improve performance. Can use the index and count overload.
- Moved logic out of the Quoted parse loop and combined single and double.
- A couple of changes mainly for consistency.
- Removed some branching per character comparison in the `Is*` functions. Logic for a-zA-Z is:
  Lowercase A-Z by ORing 0x20 (Modifying other characters is not important due to the following)
  Move all characters down so that 'a' has a value of zero
  Compare as a uint so that characters lower than 'a' (negatives) become greater positives